+++ src/js/index.ts
@@ -5,7 +5,6 @@ import * as options from './lib/options.js';
 import * as toast from './lib/toast.js';
 import * as tooltip from './lib/tooltip.js';
 import { ColorSwitch } from './components/color-switch/index.js';
-import { parseIcs } from './lib/icalparse.js';
 
 export interface Elems {
   [index:string]: HTMLElement,
@@ -470,23 +469,14 @@ function toggleAgenda() {
   flash(panel as HTMLElement, 'highlight');
 }
 
-async function updateAgenda() {
+function updateAgenda() {
   if (!OPTS.agendaUrl || OPTS.agendaUrl === chrome.i18n.getMessage('default_agenda_link')) return;
   const rootPanel = els.main.querySelector('#agendaPanel') as HTMLElement;
   if (!rootPanel) return;
   while (rootPanel.lastElementChild!.firstChild) {
     rootPanel.lastElementChild!.removeChild(rootPanel.lastElementChild!.lastChild!);
   }
-  let events;
-  try {
-    const response = await fetch(OPTS.agendaUrl);
-    const text = await response.text();
-    events = parseIcs(text);
-  } catch (e) {
-    toast.html('agenda', chrome.i18n.getMessage('bad_agenda_link'));
-    return;
-  }
-  for (const event of events.slice(0, OPTS.agendaNb)) {
+  for (const event of OPTS.events.slice(0, OPTS.agendaNb)) {
     const panel = createPanel(rootPanel.lastElementChild as HTMLElement);
     panel.firstElementChild!.textContent = (event.location) ? event.title + ' - ' + event.location : event.title;
     const p = document.createElement('p');
